+++ packages/jimp/package.json
@@ -63,6 +63,7 @@
     "@jimp/png": "0.3.5",
     "@jimp/bmp": "0.3.5",
     "@jimp/tiff": "0.3.5",
+    "@jimp/plugin-color": "0.3.5",
     "@jimp/plugin-resize": "0.3.5",
     "@jimp/plugin-blit": "0.3.5",
     "@jimp/plugin-rotate": "0.3.5",
