+++ packages/plugin-color/README.md
@@ -1,0 +1,3 @@
+# @jimp/plugin-color
+
+Bitmap manipulation to adjust the color in an image.
